Added create event test.

+++ backend/tests/setup.ts
@@ -1,16 +1,19 @@
-import { schema } from '../db/schema';
-import { createDb } from '../db/create-db';
-import DB from '../db/db';
+import { schema } from "../db/schema";
+import { createDb } from "../db/create-db";
+import DB from "../db/db";
+import knex from "knex";
+import knexConfig from "../knexfile";
 
 interface TableInfo {
-  name: string,
-};
+  name: string;
+}
 
 let db = new DB();
 
 beforeAll(async () => {
   createDb();
   await db.execAsync(schema);
+  await knex(knexConfig).migrate.latest();
 });
 
 afterAll(async () => {
@@ -22,7 +25,13 @@ beforeEach(async () => {
     "SELECT name FROM sqlite_master WHERE type='table'"
   );
 
+  const MIGRATIONS_TABLE = "knex_migrations";
+
   for (const table of tables) {
+    if(table.name === MIGRATIONS_TABLE) {
+      continue;
+    }
+
     await db.db.run(`DELETE FROM ${table.name}`);
   }
 });
